Create component list
Not quite sure if I need this.

I've created a component list in The Machine and Buttons but I never created a script to generate it! I'm a bit of a script kiddie because I don't really know shell scripting very well. I just paste together what I find online and somehow it works.