Stop component list script from running too often
The script only runs when the Xcode project file is modified. The Xcode project file is modified when files are added and removed which is when we might want to regenerate the component list. When the script runs unnecessarily, the file contents is not changed, only the modified date. Xcode will only recompile sources that include the header when the contents of the header changes.