Find weird slow effect problem
The problem is that I don't know a good way of combining effects. I think just replacing the old effect with the new one is the best way but it means that slow towers can "unfreeze" stunned units. I tried taking the min of the speedFactors but that just resulting in the slow tower constantly extending the duration of the stun effect. I think replacement is an simple enough behaviour for the player to understand and work with.