Define the depth of everything in one place
The laser beam is over the projectile towers which looks a little strange. The beam should be on top of towers but under units. But projectiles should be under tower guns and over units so tower guns have to be on top of units which means that a beam can't be under a unit and over a tower gun. I'll figure it out!