Overhaul the tower systems
From the player's perspective, nothing has changed. I realised that the difference between a projectile tower and a beam tower is just the animation and the difference between a turret tower and a splash tower is damage. I split the animation systems into projectile and beam. I split the damage systems into turret and splash. The final solution feels a little "hacky" but it is an improvement. I still need to make sure there aren't any off by one errors.

Beam towers deal damage every frame in which they have a target. This is to simplify the code a little and it kind of makes sense. A beam of energy should deal a continuous beam of damage. I just have to make sure that the damage is definitely dealt every frame. Due to floating point errors, some frames might be skipped.