Rename UnitExitDistance to UnitMoveDistance
UnitMoveDistance stores the distance between the unit and the spawn point. Towers now target the unit that is furthest from the spawn point rather than closest to the exit. Since there is only spawn point and one exit, this makes no difference. I've already assumed that  there is one spawn point and one base in other parts of the code base so I might as well make the assumption here as well.